Changing i.e. to e.g. (#150)
i.e. means "in other words", usually if you want to tell someone a more commonly used synonym to a word that they may not be familiar with.  e.g. means "for example", which seems more appropriate in the context of how i.e. is being used here.  Meter is not a synonym for length, it's an example of a unit of length.