Update the extrapolation section (#168)
* Update the extrapolation section

Updating the extrapolation section based on WI 36667 (Update the behavior of Extrapolation Mode to follow Documentation).

* Moving the inter-/extrapolation to "reading data" 

Per discussion with McKenzie, moving the interpolation and extrapolation sections out of "types" file and into reading data under read characteristics.

* Moving inter-/extrapolation out of types

Moving  inter-/extrapolation into reading data. Fixing xrefs to match the update.



* Adding more info to a link

Pointing to a specific section in a link

* Moving default table to the end 

Per Julija's comment, moving the default behavior to the bottom of the list. Also getting read reverting the wording to the original for cells in ExtrapolationMode with InterpolationMode = ContinuousNullableLeading and ContinuousNullableTrailing table.

* Shuffling rows in the table

Putting interpolation mode as the first column in the table "Default ExtrapolationMode based on InterpolationMode".

* Pointing links to a new location

Updating links in the "Interpolation and Extrapolation" section now that the content is in "Reading Data"

* Updating links for interpolation

Point the link to a new location.

* Removing a table

Removing the table "default extrapolation mode based on interpolation" as it's going to be a separate WI.

* Removing the footnote

Removing the footnote for table "**ExtrapolationMode with InterpolationMode = ContinuousNullableLeading and ContinuousNullableTrailing**"

* Removing the last sentence in "Streams" topic  per McKenzie's comment on "Interpolation and Extrapolation" section