Shared Access API Routes (#445)
* Shared Access API Routes

* Docupipe format

* notes from PR

* Update shared-access.md

* Update communities-overview.md

* Update shared-access.md

A few minor copy edits

* Update shared-access.md

I'll keep looking at this tomorrow.

When we talk about "this service brokering requests," are we talking about the SDS service?

* Examples

* Update shared-access.md

* Update shared-access.md

* Update shared-access.md

Various small edits. 

I shifted all the header levels down by 1 (for example, ## => #. In the help system, this file will as it's own topic so there's no reason to start with a subordinate heading. We should use the TOC to indicate subordinate topics.

* communities -> community

* Update shared-access.md

Co-authored-by: David Wicinas <dwicinas@osisoft.com>